ml-moderation/scripts/requirements.txt
# Python dependencies cho ML Moderation scripts
# Cài đặt: pip install -r requirements.txt

# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# MongoDB
pymongo>=4.5.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0

# API (nếu muốn chạy Flask server)
Flask>=3.0.0
Flask-CORS>=4.0.0
pyngrok>=7.0.0  # Chỉ cần nếu chạy trên Colab
